fix: исправление замечаний 3 дз
