+++ src/components/app/app.tsx
@@ -1,11 +1,42 @@
-import Routing from '../../routing';
-
+import Main from '../../pages/main/main';
+import Favorites from '../../pages/favorites/favorites';
+import Offer from '../../pages/offer/offer';
+import Login from '../../pages/login/login';
+import Error404 from '../../pages/error-404/error-404';
+import Layout from '../../components/layout/layout';
+import PrivateRoute from '../../components/private-routers/private-router';
+import PrivateLoginRoute from '../../components/private-routers/private-login-router';
+import { AppRoute } from '../../const';
+import { BrowserRouter, Routes, Route } from 'react-router-dom';
+import { HelmetProvider } from 'react-helmet-async';
 type AppProps = {
   rentalOffer: number;
 }
 
 export default function App({ rentalOffer }: AppProps): JSX.Element {
   return (
-    <Routing rentalOffer={rentalOffer} />
+    <HelmetProvider>
+      <BrowserRouter>
+        <Routes>
+          <Route path={AppRoute.Root} element={<Layout />}>
+            <Route index element={<Main rentalOffer={rentalOffer} />} />
+            <Route path={AppRoute.Favorites} element={
+              <PrivateRoute >
+                <Favorites />
+              </PrivateRoute>
+            }
+            />
+            <Route path={AppRoute.Offer} element={<Offer />} />
+            <Route path={AppRoute.Login} element={
+              <PrivateLoginRoute >
+                <Login />
+              </PrivateLoginRoute>
+            }
+            />
+          </Route>
+          <Route path='*' element={<Error404 />} />
+        </Routes>
+      </BrowserRouter>
+    </HelmetProvider>
   );
 }
