+++ src/pages/main/main.tsx
@@ -18,7 +18,7 @@ export default function Main({ rentalOffer }: MainProps): JSX.Element {
         <div className="tabs">
           <section className="locations container">
             <ul className="locations__list tabs__list">
-              {Object.values(SixCities).map((city) => (
+              {SixCities.map((city) => (
                 <li className="locations__item" key={city}>
                   <a className="locations__item-link tabs__item" href="#">
                     <span>{city}</span>
